+++ .~lock.WebDeskCalcDevLog.odt#
@@ -1,1 +1,1 @@
-,SC2-DESKTOP/SC2,SC2-Desktop,09.12.2021 22:37,file:///C:/Users/SC2/AppData/Roaming/LibreOffice/4;
+,SC2-DESKTOP/SC2,SC2-Desktop,11.12.2021 23:40,file:///C:/Users/SC2/AppData/Roaming/LibreOffice/4;
