+++ Calc/src/js/controller.js
@@ -1,10 +1,12 @@
 "use strict";
 
 import { calcResult } from "./models/calcDelegate.js";
+import { printResult } from "./views/viewPrint.js";
 
 import viewInfo from "./views/viewInfo.js";
 import viewCalc from "./views/viewCalc.js";
 import viewResult from "./views/viewResult.js";
+
 // import viewMaterials from "./views/viewMaterials.js";
 
 const calculateLadder = function () {
@@ -14,17 +16,22 @@ const calculateLadder = function () {
   calcResult(userInput, renderError, renderResult);
 };
 
+const addInputHandlers = function () {
+  viewCalc.addHandlerCalculate(calculateLadder);
+  viewCalc.addHandlerCalcEnter(calculateLadder);
+  viewCalc.stdInputFieldColor();
+};
+
 const renderLadderCalc = function () {
   viewCalc._renderData(viewCalc._renderCalculator());
-  viewCalc.addHandlerCalculate(calculateLadder);
-  viewCalc.stdInputFieldColor();
+  addInputHandlers();
 };
 
 const renderResult = function (data) {
   viewCalc._clearInputField();
   viewResult._renderData(viewResult._renderResult(data));
   viewResult.addHandlerBack(renderLadderCalc);
-  viewResult.addHandlerSave(printResult);
+  viewResult.addHandlerSave(savePrint);
   // viewMaterials._renderData(viewMaterials._renderMaterials(data));
 };
 
@@ -33,28 +40,14 @@ const renderError = function () {
   viewResult._renderData(viewResult._renderError());
 };
 
+const savePrint = function (data) {
+  printResult(data);
+};
+
 const init = function () {
   viewInfo._renderData(viewInfo._generateInfoLadder());
+  addInputHandlers();
   // viewResult._getParentEl();
-  viewCalc.addHandlerCalculate(calculateLadder);
-  viewCalc.stdInputFieldColor();
 };
 
 init();
-
-///////////////////
-
-const printResult = function (data) {
-  const printArea = window.open();
-
-  printArea.document.write("<title>Ladder Size Calculator</title>");
-  printArea.document.write('<link rel="stylesheet" href="./styles/printStyle.css" />');
-  printArea.document.write('<h2 class="subcontent-print_title">Rung Access Ladder</h2>');
-  printArea.document.write(data.innerHTML);
-
-  // Small delay before print in order to allow CSS finish loading.
-  setTimeout(function () {
-    printArea.print();
-    printArea.close();
-  }, 100);
-};
